# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.CommitRequest:
    properties:
      message:
        example: Initial commit
        type: string
    type: object
  handlers.CommitResponse:
    properties:
      commitHash:
        example: a1b2c3d4
        type: string
    type: object
...
  handlers.DeleteAccountRequest:
    properties:
      password:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  handlers.PullResponse:
    properties:
      message:
        example: Pulled changes from remote
        type: string
    type: object
...
  handlers.UpdateProfileRequest:
    properties:
      currentPassword:
        type: string
      displayName:
        type: string
      email:
        type: string
      newPassword:
        type: string
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      displayName:
        type: string
      email:
        type: string
      id:
        minimum: 1
        type: integer
      lastWorkspaceId:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/models.UserRole'
        enum:
        - admin
        - editor
        - viewer
    required:
    - email
    - id
    - role
    type: object
  models.UserRole:
    enum:
    - admin
    - editor
    - viewer
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleEditor
    - RoleViewer
...
info:
  contact: {}
  description: This is the API for Lemma markdown note taking app.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Lemma API
  version: "1.0"
paths:
...
  /profile:
    delete:
      consumes:
      - application/json
      description: Deletes the user's account and all associated data
      operationId: deleteAccount
      parameters:
      - description: Account deletion request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.DeleteAccountRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content - Account deleted successfully
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Password is incorrect
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Cannot delete the last admin account
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to delete account
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete account
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates the user's profile
      operationId: updateProfile
      parameters:
      - description: Profile update request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Current password is required to change email
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Current password is incorrect
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to update profile
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update profile
      tags:
      - users
...
  /workspaces/{workspace_name}/git/commit:
    post:
      description: Stages, commits, and pushes changes to the remote repository
      operationId: stageCommitAndPush
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      - description: Commit request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CommitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CommitResponse'
        "400":
          description: Commit message is required
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Failed to stage, commit, and push changes
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Stage, commit, and push changes
      tags:
      - git
  /workspaces/{workspace_name}/git/pull:
    post:
      description: Pulls changes from the remote repository
      operationId: pullChanges
      parameters:
      - description: Workspace name
        in: path
        name: workspace_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PullResponse'
        "500":
          description: Failed to pull changes
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Pull changes from remote
      tags:
      - git
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"